Separate constants and types from modules
