+++ src/const.js
@@ -1,0 +1,9 @@
+export const FILMS_COUNT = 17;
+
+export const FILMS_COUNT_PER_STEP = 5;
+
+export const FILMS_EXTRA_COUNT = 2;
+
+export const MAX_LENGTH_DESCRIPTION = 140;
+
+export const REGEXP_ALL_FILMS_COUNT = /\B(?=(\d{3})+(?!\d))/g;
