+++ src/utils/dom-utils.js
@@ -1,9 +1,5 @@
 import AbstractView from '../view/abstract-view';
-
-export const RenderPlace = {
-  BEFORE_END: 'beforeend',
-  AFTER_END: 'afterend',
-};
+import {RenderPlace} from '../types';
 
 export const createElement = (template) => {
   const newElement = document.createElement('div');
